# openapi/beecount-cloud-v1.yaml
openapi: 3.1.0
info:
  title: BeeCount Cloud
  description: BeeCount Cloud v1 API
  version: 0.1.0
paths:
...
  /api/v1/profile/me:
    get:
      tags:
      - profile
      summary: Get My Profile
      operationId: get_my_profile_api_v1_profile_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileOut'
      security:
      - OAuth2PasswordBearer:
        - app_write
      - OAuth2PasswordBearer:
        - web_read
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
    patch:
      tags:
      - profile
      summary: Patch My Profile
      operationId: patch_my_profile_api_v1_profile_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfilePatchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
  /api/v1/profile/avatar:
    post:
      tags:
      - profile
      summary: Upload My Avatar
      operationId: upload_my_avatar_api_v1_profile_avatar_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_my_avatar_api_v1_profile_avatar_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileAvatarUploadOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer:
        - app_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
  /api/v1/profile/avatar/{user_id}:
    get:
      tags:
      - profile
      summary: Download Avatar
      operationId: download_avatar_api_v1_profile_avatar__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            image/*:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_my_avatar_api_v1_profile_avatar_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_my_avatar_api_v1_profile_avatar_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserProfileAvatarUploadOut:
      properties:
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        avatar_version:
          type: integer
          title: Avatar Version
      type: object
      required:
      - avatar_url
      - avatar_version
      title: UserProfileAvatarUploadOut
    UserProfileOut:
      properties:
        user_id:
          type: string
          title: User Id
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        avatar_version:
          type: integer
          title: Avatar Version
      type: object
      required:
      - user_id
      - email
      - display_name
      - avatar_url
      - avatar_version
      title: UserProfileOut
    UserProfilePatchRequest:
      properties:
        display_name:
          type: string
          maxLength: 32
          minLength: 1
          title: Display Name
      type: object
      required:
      - display_name
      title: UserProfilePatchRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes:
            app_write: App client read/write scope
            web_read: Web console read scope
            web_write: Web console write scope
            ops_write: Operations and admin scope
          tokenUrl: /api/v1/auth/login